# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for SGMA (Sistema de Gestión de Mantenimiento y Activos)
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /spare-parts:
    get:
      operationId: listSpareParts
      tags: [spare-parts]
      summary: Listar repuestos
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: lowStock
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of spare parts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SparePart"
    post:
      operationId: createSparePart
      tags: [spare-parts]
      summary: Crear repuesto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SparePartInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SparePart"
  /spare-parts/{id}:
    patch:
      operationId: updateSparePart
      tags: [spare-parts]
      summary: Actualizar repuesto
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SparePartUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SparePart"
    delete:
      operationId: deleteSparePart
      tags: [spare-parts]
      summary: Eliminar repuesto
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /spare-parts/{id}/movements:
    post:
      operationId: createStockMovement
      tags: [spare-parts]
      summary: Registrar movimiento de inventario
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockMovementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SparePart"
...
  /work-centers:
    get:
      operationId: listWorkCenters
      tags: [work-centers]
      summary: Listar centros de trabajo
      responses:
        "200":
          description: List of work centers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkCenter"
    post:
      operationId: createWorkCenter
      tags: [work-centers]
      summary: Crear centro de trabajo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkCenterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkCenter"
  /work-centers/{id}:
    patch:
      operationId: updateWorkCenter
      tags: [work-centers]
      summary: Actualizar centro de trabajo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkCenterUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkCenter"
    delete:
      operationId: deleteWorkCenter
      tags: [work-centers]
      summary: Eliminar centro de trabajo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /technicians:
    get:
      operationId: listTechnicians
      tags: [technicians]
      summary: Listar personal
      responses:
        "200":
          description: List of technicians
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Technician"
    post:
      operationId: createTechnician
      tags: [technicians]
      summary: Crear personal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TechnicianInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Technician"
  /technicians/{id}:
    patch:
      operationId: updateTechnician
      tags: [technicians]
      summary: Actualizar personal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TechnicianUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Technician"
    delete:
      operationId: deleteTechnician
      tags: [technicians]
      summary: Eliminar personal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /suppliers:
    get:
      operationId: listSuppliers
      tags: [suppliers]
      summary: Listar proveedores
      responses:
        "200":
          description: List of suppliers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplier"
    post:
      operationId: createSupplier
      tags: [suppliers]
      summary: Crear proveedor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
  /suppliers/{id}:
    patch:
      operationId: updateSupplier
      tags: [suppliers]
      summary: Actualizar proveedor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    delete:
      operationId: deleteSupplier
      tags: [suppliers]
      summary: Eliminar proveedor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    SparePart:
      type: object
      required:
        - id
        - codigo
        - descripcion
        - stock
        - stockMinimo
      properties:
        id:
          type: integer
        codigo:
          type: string
        descripcion:
          type: string
        categoria:
          type: ["string", "null"]
        stock:
          type: integer
        stockMinimo:
          type: integer
        stockMaximo:
          type: ["integer", "null"]
        costoUnitario:
          type: ["number", "null"]
        ubicacionId:
          type: ["integer", "null"]
        ubicacionNombre:
          type: ["string", "null"]
    SparePartInput:
      type: object
      required:
        - codigo
        - descripcion
        - stock
        - stockMinimo
      properties:
        codigo:
          type: string
          minLength: 1
        descripcion:
          type: string
          minLength: 1
        categoria:
          type: string
        stock:
          type: integer
        stockMinimo:
          type: integer
        stockMaximo:
          type: integer
        costoUnitario:
          type: number
        ubicacionId:
          type: integer
    SparePartUpdate:
      type: object
      properties:
        codigo:
          type: string
          minLength: 1
        descripcion:
          type: string
          minLength: 1
        categoria:
          type: string
        stock:
          type: integer
        stockMinimo:
          type: integer
        stockMaximo:
          type: integer
        costoUnitario:
          type: number
        ubicacionId:
          type: integer
    StockMovementInput:
      type: object
      required:
        - tipo
        - cantidad
      properties:
        tipo:
          type: string
        cantidad:
          type: integer
        motivo:
          type: string
...
    WorkCenter:
      type: object
      required:
        - id
        - nombre
        - tipo
      properties:
        id:
          type: integer
        nombre:
          type: string
        tipo:
          type: string
        descripcion:
          type: ["string", "null"]
        responsable:
          type: ["string", "null"]
    WorkCenterInput:
      type: object
      required:
        - nombre
        - tipo
      properties:
        nombre:
          type: string
          minLength: 1
        tipo:
          type: string
        descripcion:
          type: string
        responsable:
          type: string
    WorkCenterUpdate:
      type: object
      properties:
        nombre:
          type: string
          minLength: 1
        tipo:
          type: string
        descripcion:
          type: string
        responsable:
          type: string
    Technician:
      type: object
      required:
        - id
        - nombre
        - rol
      properties:
        id:
          type: integer
        nombre:
          type: string
        rol:
          type: string
        especialidad:
          type: ["string", "null"]
        certificaciones:
          type: ["string", "null"]
        telefono:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        centroTrabajoId:
          type: ["integer", "null"]
        centroTrabajoNombre:
          type: ["string", "null"]
    TechnicianInput:
      type: object
      required:
        - nombre
        - rol
      properties:
        nombre:
          type: string
          minLength: 1
        rol:
          type: string
        especialidad:
          type: string
        certificaciones:
          type: string
        telefono:
          type: string
        email:
          type: string
        centroTrabajoId:
          type: integer
    TechnicianUpdate:
      type: object
      properties:
        nombre:
          type: string
          minLength: 1
        rol:
          type: string
        especialidad:
          type: string
        certificaciones:
          type: string
        telefono:
          type: string
        email:
          type: string
        centroTrabajoId:
          type: integer
    Supplier:
      type: object
      required:
        - id
        - nombre
        - tipo
      properties:
        id:
          type: integer
        nombre:
          type: string
        tipo:
          type: string
        contacto:
          type: ["string", "null"]
        telefono:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        calificacion:
          type: ["number", "null"]
    SupplierInput:
      type: object
      required:
        - nombre
        - tipo
      properties:
        nombre:
          type: string
          minLength: 1
        tipo:
          type: string
        contacto:
          type: string
        telefono:
          type: string
        email:
          type: string
        calificacion:
          type: number
    SupplierUpdate:
      type: object
      properties:
        nombre:
          type: string
          minLength: 1
        tipo:
          type: string
        contacto:
          type: string
        telefono:
          type: string
        email:
          type: string
        calificacion:
          type: number