# lib/api-zod/src/generated/types/sparePart.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for SGMA (Sistema de Gestión de Mantenimiento y Activos)
 * OpenAPI spec version: 0.1.0
 */

export interface SparePart {
  id: number;
  codigo: string;
  descripcion: string;
  /** @nullable */
  categoria?: string | null;
  stock: number;
  stockMinimo: number;
  /** @nullable */
  stockMaximo?: number | null;
  /** @nullable */
  costoUnitario?: number | null;
  /** @nullable */
  ubicacionId?: number | null;
  /** @nullable */
  ubicacionNombre?: string | null;
}