# lib/api-zod/src/generated/types/supplierUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for SGMA (Sistema de Gestión de Mantenimiento y Activos)
 * OpenAPI spec version: 0.1.0
 */

export interface SupplierUpdate {
  /** @minLength 1 */
  nombre?: string;
  tipo?: string;
  contacto?: string;
  telefono?: string;
  email?: string;
  calificacion?: number;
}