# lib/api-zod/src/generated/types/technicianInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for SGMA (Sistema de Gestión de Mantenimiento y Activos)
 * OpenAPI spec version: 0.1.0
 */

export interface TechnicianInput {
  /** @minLength 1 */
  nombre: string;
  rol: string;
  especialidad?: string;
  certificaciones?: string;
  telefono?: string;
  email?: string;
  centroTrabajoId?: number;
}