# lib/api-zod/src/generated/types/workCenterUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for SGMA (Sistema de Gestión de Mantenimiento y Activos)
 * OpenAPI spec version: 0.1.0
 */

export interface WorkCenterUpdate {
  /** @minLength 1 */
  nombre?: string;
  tipo?: string;
  descripcion?: string;
  responsable?: string;
}